upgraded docker client version

+++ insightedge-integration-tests/tests/src/test/scala/org/insightedge/spark/utils/InsightEdgeAdminUtils.scala
@@ -242,7 +242,7 @@ object InsightEdgeAdminUtils extends Assertions{
     }
   }
 
-  def execAndWaitFor(containerId: String, command: String): Int = {
+  def execAndWaitFor(containerId: String, command: String): Long = {
     val execCreation = docker.execCreate(containerId, Array("bash", "-c", command), DockerClient.ExecCreateParam.attachStdout(), DockerClient.ExecCreateParam.attachStderr())
     val execId = execCreation.id()
     val output = docker.execStart(execId)
